backend/mlcare_app/requirements.txt
bcrypt~=3.2.0
certifi~=2018.1.18
cffi~=1.14.3
click~=7.1.2
cmake
cryptography~=2.1.4
dataclasses~=0.6
dnspython~=1.16.0
ecdsa~=0.14.1
flask~=1.1.2
flask_bcrypt
flask_cors
flask_pymongo
idna~=2.6
itsdangerous~=1.1.0
Jinja2~=2.11.2
joblib~=0.15.1
matplotlib~=3.2.2
MarkupSafe~=1.0
MyApplication~=0.1.0
numpy~=1.18.4
pandas~=1.0.4
Pillow~=5.1.0
ply~=3.11
psutil~=5.4.2
pycparser~=2.20
pyjwt~=1.7.1
pyMongo~=3.10.1
python-dateutil~=2.6.1
python-dotenv~=0.13.0
pytz~=2018.3
PyYAML~=3.12
requests~=2.24.0
scikit-learn~=0.23.1
scipy~=1.4.1
setuptools~=39.0.1
shap==0.37.0
simplejson~=3.13.2
six~=1.14.0
sklearn~=0.0
threadpoolctl~=2.1.0
Werkzeug~=1.0.1
wheel~=0.30.0
xgboost~=1.1.0
